Draggable events on calendar

+++ frontend/controllers/RoomController.php
@@ -369,21 +369,22 @@ class RoomController extends \yii\web\Controller
             }
             $fields['Meeting']['scheduled_at'] = $datetimepicker;
 
-            $members = Yii::$app->request->post("User");
-
             if ($meeting->load($fields) && $meeting->save()) {
 
-                $oldMembers = $room->getRoomMembers()->all();
+                if ($members = Yii::$app->request->post("User")) {
 
-                foreach ($oldMembers as $member) {
-                    $member->delete();
-                }
+                    $oldMembers = $room->getRoomMembers()->all();
 
-                foreach ($members["username"] as $k => $id) {
-                    $member = new RoomMember();
-                    $member->user_profile_id = (int)$id;
-                    $member->room_id = $room->id;
-                    $member->save();
+                    foreach ($oldMembers as $member) {
+                        $member->delete();
+                    }
+
+                    foreach ($members["username"] as $k => $id) {
+                        $member = new RoomMember();
+                        $member->user_profile_id = (int)$id;
+                        $member->room_id = $room->id;
+                        $member->save();
+                    }
                 }
 
                 return Json::encode($meeting);
@@ -440,6 +441,7 @@ class RoomController extends \yii\web\Controller
 
         $events = [];
         foreach ($rooms as $key => $room) {
+            $events[$key]['meeting_id'] = $room->meeting_id;
             $events[$key]['room_id'] = $room->id;
             $events[$key]['title'] = $room->meeting->title;
             $events[$key]['duration'] = $room->meeting->duration;
